unified-storage: isolate kv package
