+++ pkg/storage/unified/resource/datastore.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/grafana/grafana/pkg/apimachinery/validation"
+	kvpkg "github.com/grafana/grafana/pkg/storage/unified/resource/kv"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/db"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/dbutil"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/rvmanager"
@@ -37,7 +38,7 @@ func mustTemplate(filename string) *template.Template {
 }
 
 const (
-	dataSection = "unified/data"
+	dataSection = kvpkg.DataSection
 	// cache
 	groupResourcesCacheKey = "group-resources"
 	// batch operations
@@ -57,8 +58,8 @@ func newDataStore(kv KV) *dataStore {
 		cache: gocache.New(time.Hour, 10*time.Minute), // 1 hour expiration, 10 minute cleanup
 	}
 
-	if sqlkv, ok := kv.(*sqlKV); ok {
-		ds.legacyDialect = sqltemplate.DialectForDriver(sqlkv.driverName)
+	if sqlkv, ok := kv.(*kvpkg.SqlKV); ok {
+		ds.legacyDialect = sqltemplate.DialectForDriver(sqlkv.DriverName)
 	}
 
 	return ds
@@ -69,18 +70,8 @@ type DataObj struct {
 	Value io.ReadCloser
 }
 
-type DataKey struct {
-	Namespace       string
-	Group           string
-	Resource        string
-	Name            string
-	ResourceVersion int64
-	Action          DataAction
-	Folder          string
-
-	// needed to maintain backwards compatibility with unified/sql
-	GUID string
-}
+// TODO: pull DataKey from kv/sqlkv into here once we don't need sql/backend backwards compatibility
+type DataKey = kvpkg.DataKey
 
 // GroupResource represents a unique group/resource combination
 type GroupResource struct {
@@ -88,60 +79,47 @@ type GroupResource struct {
 	Resource string
 }
 
-func (k DataKey) String() string {
-	return fmt.Sprintf("%s/%s/%s/%s/%d~%s~%s", k.Group, k.Resource, k.Namespace, k.Name, k.ResourceVersion, k.Action, k.Folder)
-}
-
-// Temporary while we need to support unified/sql/backend compatibility
-// Remove once we stop using RvManager in storage_backend.go
-func (k DataKey) StringWithGUID() string {
-	return fmt.Sprintf("%s/%s/%s/%s/%d~%s~%s~%s", k.Group, k.Resource, k.Namespace, k.Name, k.ResourceVersion, k.Action, k.Folder, k.GUID)
-}
-
-func (k DataKey) Equals(other DataKey) bool {
-	return k.Group == other.Group && k.Resource == other.Resource && k.Namespace == other.Namespace && k.Name == other.Name && k.ResourceVersion == other.ResourceVersion && k.Action == other.Action && k.Folder == other.Folder
-}
-
-func (k DataKey) Validate() error {
-	if k.Namespace == "" {
-		return NewValidationError("namespace", k.Namespace, ErrNamespaceRequired)
+// TODO transform this into Validate() method on DataKey once we pull that struct back here
+func validateDataKey(dataKey DataKey) error {
+	if dataKey.Namespace == "" {
+		return NewValidationError("namespace", dataKey.Namespace, ErrNamespaceRequired)
 	}
-	if k.ResourceVersion <= 0 {
-		return NewValidationError("resourceVersion", fmt.Sprintf("%d", k.ResourceVersion), ErrResourceVersionInvalid)
+	if dataKey.ResourceVersion <= 0 {
+		return NewValidationError("resourceVersion", fmt.Sprintf("%d", dataKey.ResourceVersion), ErrResourceVersionInvalid)
 	}
-	if k.Action == "" {
-		return NewValidationError("action", string(k.Action), ErrActionRequired)
+	if dataKey.Action == "" {
+		return NewValidationError("action", string(dataKey.Action), ErrActionRequired)
 	}
 
 	// Validate naming conventions for all required fields
-	if k.Namespace != clusterScopeNamespace {
-		if err := validation.IsValidNamespace(k.Namespace); err != nil {
-			return NewValidationError("namespace", k.Namespace, err[0])
+	if dataKey.Namespace != clusterScopeNamespace {
+		if err := validation.IsValidNamespace(dataKey.Namespace); err != nil {
+			return NewValidationError("namespace", dataKey.Namespace, err[0])
 		}
 	}
-	if err := validation.IsValidGroup(k.Group); err != nil {
-		return NewValidationError("group", k.Group, err[0])
+	if err := validation.IsValidGroup(dataKey.Group); err != nil {
+		return NewValidationError("group", dataKey.Group, err[0])
 	}
-	if err := validation.IsValidResource(k.Resource); err != nil {
-		return NewValidationError("resource", k.Resource, err[0])
+	if err := validation.IsValidResource(dataKey.Resource); err != nil {
+		return NewValidationError("resource", dataKey.Resource, err[0])
 	}
-	if err := validation.IsValidGrafanaName(k.Name); err != nil {
-		return NewValidationError("name", k.Name, err[0])
+	if err := validation.IsValidGrafanaName(dataKey.Name); err != nil {
+		return NewValidationError("name", dataKey.Name, err[0])
 	}
 
 	// Validate folder field if provided (optional field)
-	if k.Folder != "" {
-		if err := validation.IsValidGrafanaName(k.Folder); err != nil {
-			return NewValidationError("folder", k.Folder, err[0])
+	if dataKey.Folder != "" {
+		if err := validation.IsValidGrafanaName(dataKey.Folder); err != nil {
+			return NewValidationError("folder", dataKey.Folder, err[0])
 		}
 	}
 
 	// Validate action is one of the valid values
-	switch k.Action {
+	switch dataKey.Action {
 	case DataActionCreated, DataActionUpdated, DataActionDeleted:
 		return nil
 	default:
-		return fmt.Errorf("action '%s' is invalid: must be one of 'created', 'updated', or 'deleted'", k.Action)
+		return fmt.Errorf("action '%s' is invalid: must be one of 'created', 'updated', or 'deleted'", dataKey.Action)
 	}
 }
 
@@ -217,12 +195,10 @@ func (k GetRequestKey) Prefix() string {
 	return fmt.Sprintf("%s/%s/%s/%s/", k.Group, k.Resource, k.Namespace, k.Name)
 }
 
-type DataAction string
-
 const (
-	DataActionCreated DataAction = "created"
-	DataActionUpdated DataAction = "updated"
-	DataActionDeleted DataAction = "deleted"
+	DataActionCreated = kvpkg.DataActionCreated
+	DataActionUpdated = kvpkg.DataActionUpdated
+	DataActionDeleted = kvpkg.DataActionDeleted
 )
 
 // Keys returns all keys for a given key by iterating through the KV store
@@ -480,7 +456,7 @@ func (d *dataStore) ListResourceKeysAtRevision(ctx context.Context, options List
 }
 
 func (d *dataStore) Get(ctx context.Context, key DataKey) (io.ReadCloser, error) {
-	if err := key.Validate(); err != nil {
+	if err := validateDataKey(key); err != nil {
 		return nil, fmt.Errorf("invalid data key: %w", err)
 	}
 
@@ -495,7 +471,7 @@ func (d *dataStore) BatchGet(ctx context.Context, keys []DataKey) iter.Seq2[Data
 	return func(yield func(DataObj, error) bool) {
 		// Validate all keys first
 		for _, key := range keys {
-			if err := key.Validate(); err != nil {
+			if err := validateDataKey(key); err != nil {
 				yield(DataObj{}, fmt.Errorf("invalid data key %s: %w", key.String(), err))
 				return
 			}
@@ -545,7 +521,7 @@ func (d *dataStore) BatchGet(ctx context.Context, keys []DataKey) iter.Seq2[Data
 }
 
 func (d *dataStore) Save(ctx context.Context, key DataKey, value io.Reader) error {
-	if err := key.Validate(); err != nil {
+	if err := validateDataKey(key); err != nil {
 		return fmt.Errorf("invalid data key: %w", err)
 	}
 
@@ -569,7 +545,7 @@ func (d *dataStore) Save(ctx context.Context, key DataKey, value io.Reader) erro
 }
 
 func (d *dataStore) Delete(ctx context.Context, key DataKey) error {
-	if err := key.Validate(); err != nil {
+	if err := validateDataKey(key); err != nil {
 		return fmt.Errorf("invalid data key: %w", err)
 	}
 
@@ -617,46 +593,9 @@ func ParseKey(key string) (DataKey, error) {
 		Namespace:       parts[2],
 		Name:            parts[3],
 		ResourceVersion: rv,
-		Action:          DataAction(rvActionFolderParts[1]),
+		Action:          kvpkg.DataAction(rvActionFolderParts[1]),
 		Folder:          rvActionFolderParts[2],
 	}, nil
-}
-
-// Temporary while we need to support unified/sql/backend compatibility.
-// Remove once we stop using RvManager in storage_backend.go
-func ParseKeyWithGUID(key string) (DataKey, error) {
-	parts := strings.Split(key, "/")
-	if len(parts) != 5 {
-		return DataKey{}, fmt.Errorf("invalid key: %s", key)
-	}
-	rvActionFolderGUIDParts := strings.Split(parts[4], "~")
-	if len(rvActionFolderGUIDParts) != 4 {
-		return DataKey{}, fmt.Errorf("invalid key: %s", key)
-	}
-	rv, err := strconv.ParseInt(rvActionFolderGUIDParts[0], 10, 64)
-	if err != nil {
-		return DataKey{}, fmt.Errorf("invalid resource version '%s' in key %s: %w", rvActionFolderGUIDParts[0], key, err)
-	}
-	return DataKey{
-		Group:           parts[0],
-		Resource:        parts[1],
-		Namespace:       parts[2],
-		Name:            parts[3],
-		ResourceVersion: rv,
-		Action:          DataAction(rvActionFolderGUIDParts[1]),
-		Folder:          rvActionFolderGUIDParts[2],
-		GUID:            rvActionFolderGUIDParts[3],
-	}, nil
-}
-
-// SameResource checks if this key represents the same resource as another key.
-// It compares the identifying fields: Group, Resource, Namespace, and Name.
-// ResourceVersion, Action, and Folder are ignored as they don't identify the resource itself.
-func (k DataKey) SameResource(other DataKey) bool {
-	return k.Group == other.Group &&
-		k.Resource == other.Resource &&
-		k.Namespace == other.Namespace &&
-		k.Name == other.Name
 }
 
 // GetResourceStats returns resource stats within the data store by first discovering
@@ -885,7 +824,7 @@ func (req sqlKVLegacyUpdateHistoryRequest) Validate() error {
 //
 // TODO: remove when backwards compatibility is no longer needed.
 func (d *dataStore) applyBackwardsCompatibleChanges(ctx context.Context, tx db.Tx, event WriteEvent, key DataKey) error {
-	_, isSQLKV := d.kv.(*sqlKV)
+	_, isSQLKV := d.kv.(*kvpkg.SqlKV)
 	if !isSQLKV {
 		return nil
 	}
