+++ pkg/storage/unified/resource/datastore_test.go
@@ -9,14 +9,33 @@ import (
 	"testing"
 
 	"github.com/bwmarrin/snowflake"
+	badger "github.com/dgraph-io/badger/v4"
 	"github.com/grafana/grafana/pkg/infra/db"
 	"github.com/grafana/grafana/pkg/setting"
+	"github.com/grafana/grafana/pkg/storage/unified/resource/kv"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/db/dbimpl"
 	"github.com/grafana/grafana/pkg/util/testutil"
 	"github.com/stretchr/testify/require"
 )
 
 var node, _ = snowflake.NewNode(1)
+
+func setupTestBadgerDB(t *testing.T) *badger.DB {
+	// Create a temporary directory for the test database
+	opts := badger.DefaultOptions("").WithInMemory(true).WithLogger(nil)
+	db, err := badger.Open(opts)
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		err := db.Close()
+		require.NoError(t, err)
+	})
+	return db
+}
+
+func setupTestKV(t *testing.T) KV {
+	db := setupTestBadgerDB(t)
+	return kv.NewBadgerKV(db)
+}
 
 func setupTestDataStore(t *testing.T) *dataStore {
 	kv := setupTestKV(t)
@@ -34,7 +53,7 @@ func setupTestDataStoreSqlKv(t *testing.T) *dataStore {
 	require.NoError(t, err)
 	dbConn, err := eDB.Init(context.Background())
 	require.NoError(t, err)
-	kv, err := NewSQLKV(dbConn.SqlDB(), dbConn.DriverName())
+	kv, err := kv.NewSQLKV(dbConn.SqlDB(), dbConn.DriverName())
 	require.NoError(t, err)
 	return newDataStore(kv)
 }
@@ -602,7 +621,7 @@ func TestDataKey_Validate(t *testing.T) {
 				Resource:        "test-resource",
 				Name:            "test-name",
 				ResourceVersion: rv,
-				Action:          DataAction("unknown"),
+				Action:          kv.DataAction("unknown"),
 			},
 			expectError: true,
 			errorMsg:    "action 'unknown' is invalid",
@@ -611,7 +630,7 @@ func TestDataKey_Validate(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := tt.key.Validate()
+			err := validateDataKey(tt.key)
 			if tt.expectError {
 				require.Error(t, err)
 				if tt.errorMsg != "" {
@@ -2127,7 +2146,7 @@ func testDataStoreListResourceKeysAtRevision(t *testing.T, ctx context.Context, 
 
 		require.Len(t, resultKeys, 3) // resource1 (updated), resource2, resource4
 		names := make(map[string]int64)
-		actions := make(map[string]DataAction)
+		actions := make(map[string]kv.DataAction)
 		for _, result := range resultKeys {
 			names[result.Name] = result.ResourceVersion
 			actions[result.Name] = result.Action
@@ -2780,7 +2799,7 @@ func testDataStoreGetResourceStatsComprehensive(t *testing.T, ctx context.Contex
 					for version := 1; version <= 3; version++ {
 						rv := node.Generate().Int64()
 
-						var action DataAction
+						var action kv.DataAction
 						switch version {
 						case 1:
 							action = DataActionCreated
