+++ pkg/storage/unified/resource/eventstore.go
@@ -12,10 +12,11 @@ import (
 
 	"github.com/bwmarrin/snowflake"
 	"github.com/grafana/grafana/pkg/apimachinery/validation"
+	"github.com/grafana/grafana/pkg/storage/unified/resource/kv"
 )
 
 const (
-	eventsSection        = "unified/events"
+	eventsSection        = kv.EventsSection
 	deleteEventBatchSize = 50
 )
 
@@ -30,7 +31,7 @@ type EventKey struct {
 	Resource        string
 	Name            string
 	ResourceVersion int64
-	Action          DataAction
+	Action          kv.DataAction
 	Folder          string
 	GUID            string
 }
@@ -81,14 +82,14 @@ func (k EventKey) Validate() error {
 }
 
 type Event struct {
-	Namespace       string     `json:"namespace"`
-	Group           string     `json:"group"`
-	Resource        string     `json:"resource"`
-	Name            string     `json:"name"`
-	ResourceVersion int64      `json:"resource_version"`
-	Action          DataAction `json:"action"`
-	Folder          string     `json:"folder"`
-	PreviousRV      int64      `json:"previous_rv"`
+	Namespace       string        `json:"namespace"`
+	Group           string        `json:"group"`
+	Resource        string        `json:"resource"`
+	Name            string        `json:"name"`
+	ResourceVersion int64         `json:"resource_version"`
+	Action          kv.DataAction `json:"action"`
+	Folder          string        `json:"folder"`
+	PreviousRV      int64         `json:"previous_rv"`
 }
 
 func newEventStore(kv KV) *eventStore {
@@ -115,7 +116,7 @@ func ParseEventKey(key string) (EventKey, error) {
 		Group:           parts[2],
 		Resource:        parts[3],
 		Name:            parts[4],
-		Action:          DataAction(parts[5]),
+		Action:          kv.DataAction(parts[5]),
 		Folder:          parts[6],
 	}, nil
 }
