+++ pkg/storage/unified/resource/kv/kv_test.go
@@ -1,4 +1,4 @@
-package resource
+package kv
 
 import (
 	"context"
