+++ pkg/storage/unified/testing/kv_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/grafana/grafana/pkg/infra/db"
 	"github.com/grafana/grafana/pkg/setting"
 	"github.com/grafana/grafana/pkg/storage/unified/resource"
+	"github.com/grafana/grafana/pkg/storage/unified/resource/kv"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/db/dbimpl"
 	"github.com/grafana/grafana/pkg/tests/testsuite"
 )
@@ -42,7 +43,7 @@ func TestSQLKV(t *testing.T) {
 		require.NoError(t, err)
 		dbConn, err := eDB.Init(ctx)
 		require.NoError(t, err)
-		kv, err := resource.NewSQLKV(dbConn.SqlDB(), dbConn.DriverName())
+		kv, err := kv.NewSQLKV(dbConn.SqlDB(), dbConn.DriverName())
 		require.NoError(t, err)
 		return kv
 	}, &KVTestOptions{
